core/pyproject.toml
[tool.poetry]
name = "anfora-ragnarok"
version = "0.1.0"
description = "Anfora"
authors = ["Yabir G. <yabirg@protonmail.com>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.6"
fastapi = "^0.54.1"
tortoise-orm = "^0.15.9"
uvicorn = "^0.11.2"
pyyaml = "^5.3"
avatar-gen = {git = "https://github.com/lsdlab/avatar-gen.git"}
bcrypt = "^3.1.7"
pycryptodome = "^3.9.4"
jsonschema = "^3.2.0"
hashids = "^1.2.0"
pyjwt = "^1.7.1"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
toml = "^0.10.0"
python-box = "^4.2.2"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
requests = "^2.22.0"
asynctest = "^0.13.0"
pytest = "^5.3.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
